requirements.txt
sentinelhub>=3.9.0
scikit-learn>=1.4.0
shapely>=2.0.0
rasterio>=1.3.0
geopandas>=0.14.0
numpy>=1.26.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
requests>=2.31
beautifulsoup4>=4.12
python-docx>=1.1
pdfplumber>=0.10
asf-search>=6.7.3
hyp3-sdk>=4.1.1
pyvista>=0.43
imageio[ffmpeg]>=2.33
streamlit>=1.35
plotly>=5.20
folium>=0.16
streamlit-folium>=0.20
